move rides logic to actions
