+++ .eslintrc.json
@@ -16,6 +16,6 @@
         "sourceType": "module"
     },
     "rules": {
-        "max-len": ["error", { "code": 170 }]
+        "max-len": ["error", { "code": 256 }]
     }
 }
