+++ src/actions/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+  rides: require('./rides'),
+};
