+++ src/routes/rides.js
@@ -3,6 +3,8 @@ const app = module.exports = require('express')();
 const bodyParser = require('body-parser');
 const jsonParser = bodyParser.json();
 
+const {insertRide, rideList, rideById} = require('./../actions').rides;
+
 module.exports = (db) => {
   /**
      * @api {post} /rides Insert New Rides to DB
@@ -98,71 +100,16 @@ module.exports = (db) => {
     });
     // END: Dirty Code #2
 
-    const startLatitude = Number(req.body.start_lat);
-    const startLongitude = Number(req.body.start_long);
-    const endLatitude = Number(req.body.end_lat);
-    const endLongitude = Number(req.body.end_long);
-    const riderName = req.body.rider_name;
-    const driverName = req.body.driver_name;
-    const driverVehicle = req.body.driver_vehicle;
-
-    if (startLatitude < -90 || startLatitude > 90 || startLongitude < -180 || startLongitude > 180) {
-      return res.status(400).send({
-        error_code: 'VALIDATION_ERROR',
-        message: 'Start latitude and longitude must be between -90 - 90 and -180 to 180 degrees respectively',
-      });
-    }
-
-    if (endLatitude < -90 || endLatitude > 90 || endLongitude < -180 || endLongitude > 180) {
-      return res.status(400).send({
-        error_code: 'VALIDATION_ERROR',
-        message: 'End latitude and longitude must be between -90 - 90 and -180 to 180 degrees respectively',
-      });
-    }
-
-    if (typeof riderName !== 'string' || riderName.length < 1) {
-      return res.status(400).send({
-        error_code: 'VALIDATION_ERROR',
-        message: 'Rider name must be a non empty string',
-      });
-    }
-
-    if (typeof driverName !== 'string' || driverName.length < 1) {
-      return res.status(400).send({
-        error_code: 'VALIDATION_ERROR',
-        message: 'Rider name must be a non empty string',
-      });
-    }
-
-    if (typeof driverVehicle !== 'string' || driverVehicle.length < 1) {
-      return res.status(400).send({
-        error_code: 'VALIDATION_ERROR',
-        message: 'Rider name must be a non empty string',
-      });
-    }
-
-    const values = [req.body.start_lat, req.body.start_long, req.body.end_lat, req.body.end_long, req.body.rider_name, req.body.driver_name, req.body.driver_vehicle];
-
-    db.run('INSERT INTO Rides(startLat, startLong, endLat, endLong, riderName, driverName, driverVehicle) VALUES (?, ?, ?, ?, ?, ?, ?)', values, function(err) {
-      if (err) {
-        return res.status(500).send({
-          error_code: 'SERVER_ERROR',
-          message: 'Unknown error',
-        });
+    insertRide(body, db, function(result) {
+      if (result.error_code == undefined) {
+        res.send(result);
+      } else if (result.error_code == 'SERVER_ERROR') {
+        res.status(500).send(result);
+      } else if (result.error_code == 'VALIDATION_ERROR') {
+        res.status(400).send(result);
+      } else {
+        res.status(404).send(result);
       }
-
-      console.log(db.lastID);
-
-      db.all('SELECT * FROM Rides WHERE rideID = ?', db.lastID, function(err, rows) {
-        if (err) {
-          return res.status(500).send({
-            error_code: 'SERVER_ERROR',
-            message: 'Unknown error',
-          });
-        }
-
-        res.send(rows);
-      });
     });
   });
 
@@ -256,61 +203,13 @@ module.exports = (db) => {
     });
     // END: Dirty Code #2
 
-    let limit = 100;
-    if (req.body.limit != undefined) {
-      limit = req.body.limit;
-    }
-    let page = 1;
-    if (req.body.page != undefined) {
-      page = req.body.page;
-    }
-    let totalPage = 1;
-    let totalRecord = 0;
-    const offset = (page-1) * limit;
-
-    db.get('SELECT count(rideID) total_record FROM Rides', function(err, result) {
-      if (err) {
-        return res.status(500).send({
-          error_code: 'SERVER_ERROR',
-          message: 'Unknown error',
-        });
-      }
-
-      totalRecord = result.total_record;
-      if (totalRecord > 0) {
-        totalPage = Math.ceil(totalRecord/limit);
-
-        if (page > totalPage) {
-          return res.send({
-            records: [],
-            page: page,
-            limit: limit,
-            total_page: totalPage,
-            total_record: totalRecord,
-          });
-        } else {
-          db.all('SELECT * FROM Rides LIMIT '+limit+' OFFSET '+offset, function(err, rows) {
-            if (err) {
-              return res.status(500).send({
-                error_code: 'SERVER_ERROR',
-                message: 'Unknown error',
-              });
-            }
-
-            res.send({
-              records: rows,
-              page: page,
-              limit: limit,
-              total_page: totalPage,
-              total_record: totalRecord,
-            });
-          });
-        }
+    rideList(req, db, function(result) {
+      if (result.error_code == undefined) {
+        res.send(result);
+      } else if (result.error_code == 'SERVER_ERROR') {
+        res.status(500).send(result);
       } else {
-        res.status(404).send({
-          error_code: 'RIDES_NOT_FOUND_ERROR',
-          message: 'Could not find any rides',
-        });
+        res.status(404).send(result);
       }
     });
   });
@@ -388,22 +287,14 @@ module.exports = (db) => {
     });
     // END: Dirty Code #2
 
-    db.get(`SELECT * FROM Rides WHERE rideID='${req.params.id}'`, function(err, record) {
-      if (err) {
-        return res.status(500).send({
-          error_code: 'SERVER_ERROR',
-          message: 'Unknown error',
-        });
+    rideById(req, db, function(result) {
+      if (result.error_code == undefined) {
+        res.send(result);
+      } else if (result.error_code == 'SERVER_ERROR') {
+        res.status(500).send(result);
+      } else {
+        res.status(404).send(result);
       }
-
-      if (record == null) {
-        return res.status(404).send({
-          error_code: 'RIDES_NOT_FOUND_ERROR',
-          message: 'Could not find any rides',
-        });
-      }
-
-      res.send(record);
     });
   });
